==> processing/03_children_per_county.ipynb <==
## Children Household Count Per County File Processing

In this file, we will be making a dataframe that includes the number of children in all of the households in each county of the US.

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import geopandas as gpd

### Importing county level shape file

This shapefile of US counties was also found online. This file provides the polygon geometries of each county in the US; this will allow us to make visualizations of the US with breakdowns by county. In this case, we are making this processing file to allow us make a map of US with children count by county.

df_counties = gpd.read_file('/hpc/group/codeplus22-vis/celine_data/source_files/county_shapefiles/counties.shp')

df_counties['county_fips'] = df_counties['STATEFP'] + df_counties['COUNTYFP']
df_counties

### Importing infousa data

This preprocessed infousa household data contains the children counts of households as well as the zipcode and county fips of each household. This file also includes the transformed latitude and longitude coordinates.

In this dataset, the county_fips column represents each individual county; this is the column we will be grouping the children count by. 

df_hh = pd.read_parquet('/hpc/group/codeplus22-vis/celine_data/zip_00_99.parquet')
df_hh

df_hh = df_hh[['county_fips', 'state', 'child_num', 'has_child', 'age_code']]
df_hh

### Groupby

Here, we are taking the infousa preprocessed file and calculating the number of children per county. We are doing so using the ```groupby``` function, which groups the county_fips together by summing the corresponding ```child_num``` values. The resulting dataframe is giving the calculated number of children per county_fips.

df_child_count = df_hh.groupby('county_fips')['child_num'].sum().reset_index()
df_child_count

### Merging child number counts with county geometries

Now that we have the child_num calculated for each ```county_fips```, we are merging the county geometry shapefile with the child count dataframe on the ```county_fips``` column. Afterwards, we will select on the columns we want to keep in this dataframe, before exporting this as a parquet file to be used in future visualizations.

df_final = df_child_count.merge(df_counties, on = ['county_fips'], how = 'left')

df_final

df_final = df_final[['STATEFP', 'NAME', 'county_fips','child_num', 'geometry']]
df_final

df_final.rename(columns = {'STATEFP': 'state', 'NAME':'county'}, inplace = True)
df_final

### Converting GeoDataFrame

In order to export this dataframe as a shapefile, the pandas dataframe must be converted to a GeoDataFrame using the ```gpd.GeoDataFrame()``` function. 

gpd_df = gpd.GeoDataFrame(df_final)
gpd_df

gpd_df.to_file('/hpc/group/codeplus22-vis/celine_data/children_count_by_county.shp')

df =gpd.read_file('/hpc/group/codeplus22-vis/celine_data/children_count_by_county.shp')
df
