==> processing/04_risk_per_tank.ipynb <==
# Classifying tanks by whether or not tanks lie on floodplains (using spatial joins), and merging with National Risk Index Data

### Import statements

import pandas as pd
import geopandas as gpd

### Reading AST Data

df_tanks = gpd.read_file('/hpc/group/codeplus22-vis/celine_data/ast_master.shp')
df_tanks.head(n=3)

### Using ```.sjoin()``` to classify tanks by whether or not they are on a floodplain

#### Reading floodplain data from the Federal Emergency Management Agency (FEMA)
Then filtering for only the column we need (geometry) to find which tanks lie on a floodplain.

df_floodplains = gpd.read_file('/hpc/group/codeplus22-vis/celine_data/source_files/nat_hazard_data/NFHL/NFHL_Key_Layers.gdb')
df_floodplains

df_floodplains = df_floodplains[['geometry']]
df_floodplains

#### Using the GeoPandas buffer function
This way, tanks within 200 meters from either side of the floodplain will be marked as near a floodplain. As The floodplain data is given in linestring or multilinestring geometries, so buffering it provides a more generalized understanding of the tanks that are near areas of flooding risk, not only the ones directly in the line of risk.

In order to buffer the geometries by 200 meters, it is necessary to convert the coordinate system of the dataframe to EPSG 3857, as the unit of measurement for this coordinate system is the meter. The final coordinate system conversion sets the floodplain dataframe to the same coordinate system as the tank dataframe. This consistency is key in the next few steps. 

%%time
df_floodplains = df_floodplains.to_crs("EPSG:3857")
df_floodplains = df_floodplains.buffer(200)
df_floodplains= df_floodplains.to_crs("EPSG:4326")
df_floodplains

The buffer function transforms the geopandas dataframe into a geoseries, but to find which tanks lie on or near floodplains, it is necessary to convert this geoseries into a GeoDataFrame.

gdf_floodplains = gpd.GeoDataFrame(df_floodplains)
gdf_floodplains.rename(columns = {0: 'geometry'}, inplace = True)
gdf_floodplains

#### Finding the tanks that lie on/near floodplains using the sjoin() function

The GeoPandas ```.sjoin()``` function performs a spatial join of two GeoDataFrames. In this case, the predicate parameter input 'intersects' means that the output is a new GeoDataFrame containing only the rows in one GeoDataFrame whose geometries intersected those in the other GeoDataFrame. In other words, function outputs a GeoDataFrame containing only the tanks that lie within 200 meters of a floodplain.

%%time
df_intersect = gpd.sjoin(df_tanks, gdf_floodplains, how='inner', predicate='intersects')
df_intersect.head()

Drop tanks with the same latitude and longitude, which are therefore duplicates:

df_intersect = df_intersect.drop_duplicates(subset = ['lat_t_4326', 'lon_t_4326'])
df_intersect.head()

#### Using ```df_intersect``` to classify the tanks in the original dataset

idx = list(df_intersect.index.values)

%%time
df_tanks['on_floodplain'] = 0

for num in idx:
    df_tanks['on_floodplain'].iloc[num] = 1
    
df_tanks

Now, our original tanks dataframe, ```df_tanks``` has a column indicating whether or not that tank is near a floodplain.

### Reading National Risk Index Data, taken from FEMA
We want to classify each tank by its risk from a variety of natural hazards. To do this, we read in NRI data, filter for only the columns we want, as stipulated by our researcher, and rename them for standardization purposes.

df_nri = gpd.read_file('/hpc/group/codeplus22-vis/celine_data/source_files/nat_hazard_data/NRI_Data/NRI_GDB_Counties/NRI_GDB_Counties.gdb')
df_nri.head()

df_nri = df_nri[['STCOFIPS', 'ERQK_RISKS', 'SWND_RISKS', 'HRCN_RISKS', 'TRND_RISKS', 'CFLD_RISKS', 'RFLD_RISKS']]
df_nri.rename(columns = {'STCOFIPS': 'county', 'ERQK_RISKS': 'erqk_risks', 'SWND_RISKS': 'swnd_risks', 'HRCN_RISKS': 'hrcn_risks', 
                         'TRND_RISKS': 'trnd_risks', 'CFLD_RISKS': 'cfld_risks', 'RFLD_RISKS': 'rfld_risks'}, inplace = True)
df_nri

Then, we merge this ```df_nri``` dataframe with our ```df_tanks``` dataframe, based on the ```county``` column. Therefore, tanks are associated with risks from each natural hazard based on the county in which it is located.

### Merging AST and NRI data using pandas' ```.merge()```

df_tank_risks = df_tanks.merge(df_nri, on = 'county', how = 'left')
df_tank_risks.head()

### Calculating average risk and handling NaN values
We noticed the NRI data had a significant amount of NaN values, indicated that there is no information for that cell. Dropping all the rows with NaN values would eliminate two-thirds of our data. However, we noticed that NaN values were generally inputted for counties that had little to no risk for that specific natural hazards- counties in the center of the US had NaN values for the coastal flooding risk, for example. Therefore, after discussion with our researcher, we decided to calculate the average risk using ```0``` as the value for all NaN values, but then fill all the NaN values as ```-1```, an implausible number, to indicate in our visualizations that these values were unavailable.

To do this, we made a copy of the original ```df_tank_risks```, which had all the tank information, with natural hazard risks associated to each tank. Then, we filled NaN values of the copy of that dataframe, ```df_tank_risks_calc```, with the value ```0```. This is the dataframe we used to calculate the average risk for each tank, by adding all the risk indices and dividing it by the number of natural hazards (6). We also calculated ```adj_risk```, which is the average risk for the tank, adjusted for whether or not that tank lies near a floodplain. For this column, we added five points to the ```avg_risk``` if the tank was near a floodplain, using the ```on_floodplain``` column.

We also dropped all columns other than ```avg_risk``` and ```adj_risk```, because we will then merge this dataframe with the original ```df_tank_risks``` dataframe so that for each tank, we have risk indices for each individual natural hazards, along with these computed risks.

df_tank_risks_calc = df_tank_risks
df_tank_risks_calc = df_tank_risks_calc.fillna(0)
df_tank_risks_calc['avg_risk'] = (df_tank_risks_calc['erqk_risks'] + df_tank_risks_calc['swnd_risks'] + 
                               df_tank_risks_calc['hrcn_risks'] + df_tank_risks_calc['trnd_risks'] + 
                               df_tank_risks_calc['cfld_risks'] + df_tank_risks_calc['rfld_risks']) / 6
df_tank_risks_calc['adj_risk'] = df_tank_risks_calc['avg_risk'] + (5 * df_tank_risks_calc['on_floodplain'])
df_tank_risks_calc = df_tank_risks_calc[['avg_risk', 'adj_risk']]
df_tank_risks_calc

tank_risks_merged = pd.merge(df_tank_risks, df_tank_risks_calc, left_index = True, right_index = True)
tank_risks_merged

Finally, we fill in the NaN values for the risk data with -1, as discussed previously, and save this dataframe as a shapefile.

values = {"erqk_risks": -1, "swnd_risks": -1, "hrcn_risks": -1, "trnd_risks": -1, "cfld_risks": -1, "rfld_risks": -1}
tank_risks_merged = tank_risks_merged.fillna(value=values)
tank_risks_merged

tank_risks_merged.to_file('/hpc/group/codeplus22-vis/celine_data/tanks_risk_score.shp')
