==> processing/07_hh_count_by_county.ipynb <==
# Calculating the number of households within certain distances from tanks for each county using ```.sjoin()```

### Import statements

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import geopandas as gpd
import dask
from dask import dataframe as dd
import dask_geopandas

### Reading county shapefile dataframe
To count the number of households within five miles of a storage tank for each county, we use GeoPandas' ```.sjoin()``` method. Using this method, we will perform a spatial join between each county's geometry and the dataframe including Point geometries for each household in the US. For this, we need a dataframe with geometries for all counties in the US - which we took from the United States Census Bureau's Cartographic Boundary Files (available [here](https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html)). Then, we filter to exclude counties from Alaska, Hawaii, Puerto Rico, Virgin Islands, American Samoa, Guam, Northern Marian Islands, as there are no tanks in those regions in the AST dataset. We also drop unnecessary columns.

df_counties = gpd.read_file('/hpc/group/codeplus22-vis/celine_data/source_files/county_shapefiles/counties.shp')
df_counties = df_counties[df_counties['STATEFP'] != '02']
df_counties = df_counties[df_counties['STATEFP'] != '15']
df_counties = df_counties[df_counties['STATEFP'] != '72']
df_counties = df_counties[df_counties['STATEFP'] != '78']
df_counties = df_counties[df_counties['STATEFP'] != '60']
df_counties = df_counties[df_counties['STATEFP'] != '66']
df_counties = df_counties[df_counties['STATEFP'] != '69']
df_counties = df_counties[['NAME', 'geometry']]
df_counties.rename(columns = {'NAME': 'county'}, inplace = True)
df_counties

### Reading InfoUSA Data (pre-processed)
This InfoUSA data was preprocessed to include the distance between each household and the nearest tank. We need this information to filter for only households within five miles of a storage tank, which we do below.

df_hh = pd.read_parquet('/hpc/group/codeplus22-vis/celine_data/dist_all_hh_with_children.parquet')
df_hh.head()



We also only keep the ```lat_h_4326``` and ```lon_h_4326``` columns because these are the coordinates for each household we will use to create the Point geometries for each household in the households GeoDataFrame. We use the coordinates in EPSG 4326 rather than the coordinates in EPSG 3857 because the geometries in the ```df_counties``` GeoDataFrame are in EPSG 4326. The coordinate system must remain consistent in order to get accurate results when performing the spatial join with ```.sjoin()```.

df_hh = df_hh[['lat_h_4326', 'lon_h_4326', 'distance_category']]
df_hh = df_hh[df_hh['distance_category'] != 4]
df_hh.head()

### Use Dask to transform pandas dataframe to a geopandas dataframe
To perform a spatial join between two dataframes, each of these dataframes must be GeoDataFrames- thus, we must convert our ```df_hh``` dataframe to a GeoDataFrame. However, as this dataframe has more than 17 million rows, converting it without using Dask is not time-efficient. Hence, we turned to Dask, an open-source Python library for parallel computing. It allows us to efficiently execute the transformation of our dataframe to a GeoDataFrame, even when working with over 17 million rows. 

To use Dask, we first converted our dataframe to a Dask dataframe, using Dask's ```.from_pandas()``` method. This method takes in our pandas dataframe along with the ```npartitions``` parameter, which is used to specify the number of 'sections' the dask dataframe will be split into.

df_dask = dd.from_pandas(df_hh, npartitions = 500)

Then, we specify what manipulation to the dask dataframe ```df_dask``` to compute. In this case, we use Dask Geopandas' ```.points_from_xy()``` method to convert the pandas dask dataframe into a geopandas dask dataframe.

%%time
df_dask['geometry'] = dask_geopandas.points_from_xy(df_dask, 'lon_h_4326', 'lat_h_4326')

After, we convert the dask geodataframe into a geopandas dataframe:

%%time
gdf = dask_geopandas.from_dask_dataframe(df_dask)

Calling compute puts all the above code into action. Dask executes each set of commands on each partition, as specified above. This returns GeoDataFrame ```gdf_hh```, with over 17 million rows, in around five seconds.

%%time
gdf_hh = gdf.compute()

gdf_hh = gdf_hh[['distance_category', 'geometry']]
gdf_hh.head()

### Computing number of households within 0.5 mi of a storage tank in each county
Now that we have our two GeoDataFrames, we ```.sjoin()``` to find which households are within 0.5 miles from a tank in each county.

#### Filtering the household data for only households within 0.5 mi of a storage tank
First, we created a new GeoDataFrame ```gdf_hh_half_mi``` that included only the households within 0.5 miles of a storage tank.

gdf_hh_half_mi = gdf_hh[gdf_hh['distance_category'] == 1]
gdf_hh_half_mi.head()

#### Adding column in county dataframe with the number of households within 0.5mi of a tank
Then, we used ```.sjoin()``` to find the number of households within 0.5 miles of a tank for each county.

Calling ```.sjoin()``` and passing in two GeoDataFrames will return a new GeoDataFrame that only includes the observations with geometries that are the intersections of the two original GeoDataFrames. In our case, passing in a GeoDataFrame with the geometry for Harris County and a GeoDataFrame with all the households within 0.5 miles of a tank to the ```.sjoin()``` method returns a new GeoDataFrame with all the households within 0.5 miles of a tank in Harris County, as it returns all the Point geometries that intersects the Harris county Polygon geometry. Computing the length of that new dataframe through ```len()``` returns the number of households within 0.5 miles in Harris county, then.

However, we need to do this for all counties in the US, so we use for loop. This loop iterates every row of ```df_counties``` finds the intersection between that county and the households within 0.5 miles of a tank (```gdf_hh_half_mi```), computes the length of that ```intersect_df```, and adds that number to a new column in ```df_counties```. We add the column to ```df_counties``` specifically, because then we can use this exact dataframe to plot a map of the US where each county is colored by the number of households within five miles of a storage tank, as it contains all the geometries for each county required for a GeoViews visualization.

%%time

df_counties['hh_half_mi'] = 0
df_counties_temp = df_counties

for i in range(0, len(df_counties)):
    county = df_counties_temp.head(n=1)
    intersect_df = gpd.sjoin(county, gdf_hh_half_mi, how='inner', predicate='intersects')
    df_counties_temp = df_counties_temp.iloc[1:, :]
    num = len(intersect_df)
    # print(num)
    
    # adding num to df_counties column
    df_counties['hh_half_mi'].iloc[i] = num

df_counties

df_counties = gpd.read_file('/hpc/group/codeplus22-vis/celine_data/hh_per_county.shp')
df_counties

### Computing number of households within 1 mi of a storage tank in each county
Following the same steps as described above, except filtering for households within one mile of a storage tank.

#### Filtering the household data for only households within 1 mi of a storage tank

gdf_hh_one_mi = gdf_hh[gdf_hh['distance_category'] == 2]
gdf_hh_one_mi

#### Adding column in county dataframe with the number of households within 1 mi of a tank
Takes around 30min

%%time

df_counties['hh_one_mi'] = 0
df_counties_temp = df_counties

for i in range(0, len(df_counties)):
    county = df_counties_temp.head(n=1)
    intersect_df = gpd.sjoin(county, gdf_hh_one_mi, how='inner', predicate='intersects')
    df_counties_temp = df_counties_temp.iloc[1:, :]
    num = len(intersect_df)
    # print(num)
    
    # adding num to df_counties column
    df_counties['hh_one_mi'].iloc[i] = num

df_counties

df_counties.to_file('/hpc/group/codeplus22-vis/infousa_copy/hh_num_counties_all_final.shp')

df_counties = gpd.read_file('/hpc/group/codeplus22-vis/infousa_copy/hh_num_counties_all_final.shp')
df_counties

### Computing number of households within 5mi of a storage tank in each county
Following the same steps as described above, except filtering for households within five miles of a storage tank. Because the households GeoDataFrame has over 14 million rows, this process takes quite a while.

#### Filtering the household data for only households within 5 mi of a storage tank

gdf_hh_five_mi = gdf_hh[gdf_hh['distance_category'] == 3]
gdf_hh_five_mi

#### Adding column in county dataframe with the number of households within 5km of a tank
Takes around 4 hours.

%%time 

df_counties['hh_five_mi'] = 0
df_counties_temp = df_counties

for i in range(0, len(df_counties)):
    county = df_counties_temp.head(n=1)
    intersect_df = gpd.sjoin(county, gdf_hh_five_mi, how='inner', predicate='intersects')
    df_counties_temp = df_counties_temp.iloc[1:, :]
    num = len(intersect_df)
    
    # adding num to df_counties column 
    df_counties['hh_five_mi'].iloc[i] = num

df_counties

df_counties['HH_total'] = df_counties['HH_half_mi'] + df_counties['HH_one_mi'] + df_counties['HH_five_mi']
df_counties

### Exporting it as a shapefile

df_counties.to_file('/hpc/group/codeplus22-vis/celine_data/hh_per_county.shp')

df = gpd.read_file('/hpc/group/codeplus22-vis/celine_data/hh_per_county.shp')
df
