==> processing/08_natural_hazards_processing.ipynb <==
# Natural Hazard Risk Processing File

Processing NRI risk data with the household and tank risk data. 

### Importing libraries

import pandas as pd
import geopandas as gpd
import numpy as np

### Importing data

This is a preprocessed file with tank and household distances, transformed latitude longitude coordinates, as well as all of the risk data for 6 natural hazards (earthquake, strong winds, hurricanes, tornadoes, cold floods, and riverine floods).

df=pd.read_parquet('/hpc/group/codeplus22-vis/celine_data/dist_all_hh_with_children.parquet')
df

df.to_parquet('/hpc/group/codeplus22-vis/celine_data/hh_tank_risk.parquet')


#### Here we are just re-reading the file in to test if it exported as a parquet correctly

df = pd.read_parquet('/hpc/group/codeplus22-vis/celine_data/hh_tank_risk.parquet')
df

## Processing dataframe for individual natural hazards

Since we want to plot all of the risk dashboards separately, are now breaking down this dataframe, one for each type of natural hazard.

df_erqk = df.drop(['swnd_risks', 'hrcn_risks', 'trnd_risks', 'cfld_risks', 'rfld_risks', 'avg_risk'], axis = 1)
df_erqk = df_erqk[['is_elderly', 'distance_mi', 'erqk_risks', 'lat_3857', 'lon_3857']]
df_erqk.rename(columns = {'erqk_risks': 'earthquake_risk'}, inplace = True)

df_erqk

df_erqk.to_parquet('/hpc/group/codeplus22-vis/celine_data/earthquake_risk.parquet')

### Processing dataframe for strong wind risk

We are doing the same thing as for the earthquake file, this time dropping all of the other risks except for strong wind risk before exporting it to a parquet file to be used in visualizations.

df_swnd = df.drop(['erqk_risks', 'hrcn_risks', 'trnd_risks', 'cfld_risks', 'rfld_risks', 'avg_risk'], axis = 1)
df_swnd = df_swnd[['is_elderly', 'distance_mi', 'swnd_risks', 'lat_3857', 'lon_3857']]
df_swnd.rename(columns = {'swnd_risks': 'strong_wind_risk'}, inplace = True)


df_swnd

df_swnd.to_parquet('/hpc/group/codeplus22-vis/celine_data/strong_wind_risk.parquet')

### Processing dataframe for hurricane risk

Once again, we will be doing the same steps as the ones above, except we are keeping the column for hurricane risk. Every time we make a new dataframe, we will export it as a parquet file.

df_hrcn = df.drop(['erqk_risks','swnd_risks', 'trnd_risks', 'cfld_risks', 'rfld_risks', 'avg_risk'], axis = 1)
df_hrcn = df_hrcn[['is_elderly', 'distance_mi', 'hrcn_risks', 'lat_3857', 'lon_3857']]
df_hrcn.rename(columns = {'hrcn_risks': 'hurricane_risk'}, inplace = True)

df_hrcn

df_hrcn.to_parquet('/hpc/group/codeplus22-vis/celine_data/hurricane_risk.parquet')

### Processing dataframe for tornado risk

df_trnd = df.drop(['erqk_risks','swnd_risks', 'hrcn_risks', 'cfld_risks', 'rfld_risks', 'avg_risk'], axis = 1)
df_trnd = df_trnd[['is_elderly', 'distance_mi','trnd_risks', 'lat_3857', 'lon_3857']]
df_trnd.rename(columns = {'trnd_risks': 'tornado_risk'}, inplace = True)

df_trnd

df_trnd.to_parquet('/hpc/group/codeplus22-vis/celine_data/tornado_risk.parquet')

### Processing dataframe for coastal flood risk

df_cfld = df.drop(['erqk_risks','swnd_risks', 'trnd_risks', 'hrcn_risks', 'rfld_risks', 'avg_risk'], axis = 1)
df_cfld = df_cfld[['is_elderly', 'distance_mi','cfld_risks', 'lat_3857', 'lon_3857']]
df_cfld.rename(columns = {'cfld_risks': 'coastal_flood_risk'}, inplace = True)

df_cfld

df_cfld.to_parquet('/hpc/group/codeplus22-vis/celine_data/coast_flood_risk.parquet')

### Processing dataframe for riverine flood risk

df_rfld = df.drop(['erqk_risks','swnd_risks', 'trnd_risks', 'hrcn_risks', 'cfld_risks', 'avg_risk'], axis = 1)
df_rfld = df_rfld[[ 'is_elderly', 'distance_mi', 'rfld_risks', 'lat_3857', 'lon_3857']]
df_rfld.rename(columns = {'rfld_risks': 'riverine_flood_risk'}, inplace = True)

df_rfld

df_rfld.to_parquet('/hpc/group/codeplus22-vis/celine_data/riverine_flood_risk.parquet')

### Processing dataframe for average flood risk

df_avg = df.drop(['erqk_risks','swnd_risks', 'trnd_risks', 'hrcn_risks', 'cfld_risks', 'rfld_risks'], axis = 1)
df_avg = df_avg[[ 'is_elderly', 'distance_mi', 'avg_risk', 'lat_3857', 'lon_3857']]
df_avg.rename(columns = {'avg_risk': 'average_risk'}, inplace = True)

df_avg

df_avg.to_parquet('/hpc/group/codeplus22-vis/celine_data/average_risk.parquet')
